Initial commit of eventrewrite.d
eventrewrite.d is a scrip that will scan pymap.d
for any devices it support and then scans the
input system to see if any of those devices exist.

If it exists it will rewrite the events it recieves
and pipe them through a new uinput device.

The ForwardDevice was also abstracted into it's own
module.